[css-grid] Always consider baseline shim for the minimum contribution
Before this patch, the shim used for baseline alignment is not taken
into consideration when calculating the minimum contribution of a grid
item if its preferred size is auto, its minumum size is auto, and the
max track sizing function is a fixed small value. In this case, the
auto minimum size can produce a smaller track than if it was 0.

The specification is not clear about the interaction of minimum
contributions and baseline shims, but the above doesn't make sense.

BUG=934535

TEST=external/wpt/css/css-grid/layout-algorithm/grid-minimum-contribution-baseline-shim.html

Change-Id: Iafd194ab7e5656b5018e0e36b915ba183bdedd19